refactor: web specific context
