+++ internal/web/util/gitops.go
@@ -5,12 +5,13 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/glasskube/glasskube/internal/clicontext"
+	webcontext "github.com/glasskube/glasskube/internal/web/context"
+
 	"github.com/glasskube/glasskube/internal/telemetry/annotations"
 )
 
 func IsGitopsModeEnabled(req *http.Request) bool {
-	coreListers := clicontext.CoreListersFromContext(req.Context())
+	coreListers := webcontext.CoreListersFromContext(req.Context())
 	if coreListers != nil && coreListers.NamespaceLister != nil {
 		if ns, err := (*coreListers.NamespaceLister).Get("glasskube-system"); err != nil {
 			fmt.Fprintf(os.Stderr, "failed to determine GitOps mode: %v\n", err)
